mirror

+++ src/qppcad/ws_item/node_book/qnode_socket.cpp
@@ -20,7 +20,7 @@ qnode_socket_t::qnode_socket_t(QGraphicsItem *parent,
 
   QPainterPath p;
 
-  p.addRect(0, 0, 2*m_socket_radius, 2*m_socket_radius);
+  p.addRect(0, 0, 2 * m_socket_radius, 2 * m_socket_radius);
   setPath(p);
   setPen(QPen(m_socket_color, 2));
   setBrush(Qt::black);
@@ -33,6 +33,26 @@ qnode_socket_t::qnode_socket_t(QGraphicsItem *parent,
 
 qnode_socket_t::~qnode_socket_t() {
 
+}
+
+opt<size_t> qnode_socket_t::connections_count() {
+
+  if (!m_node) return std::nullopt;
+  if (!m_node->m_sflow_node) return std::nullopt;
+
+  size_t ret_con_count{0};
+
+  for (auto con : m_node->m_scene->m_connections)
+    if (con && con->m_inp_socket && con->m_out_socket &&
+        (con->m_inp_socket == this || con->m_inp_socket == this))
+          ret_con_count++;
+
+  return opt<size_t>(ret_con_count);
+
+}
+
+int qnode_socket_t::type() const {
+  return Type;
 }
 
 void qnode_socket_t::mousePressEvent(QGraphicsSceneMouseEvent *event) {
@@ -79,35 +99,42 @@ void qnode_socket_t::mouseReleaseEvent(QGraphicsSceneMouseEvent *event) {
 
   if (m_connection) {
 
-      QGraphicsItem *item =
-          scene()->itemAt(event->scenePos(), scene()->views().first()->transform());
+      QList<QGraphicsItem *> items = scene()->items(event->scenePos());
+      QGraphicsItem *item{nullptr};
+
+      for (auto _item : items) {
+          if (_item && _item->type() == qnode_socket_t::Type)
+            item = _item;
+        }
 
       bool delete_item = true;
 
-      if (item && item != this) {
+      if (item) {
 
-          auto item_c = qgraphicsitem_cast<qnode_socket_t*>(item);
+          qnode_socket_t* item_c = qgraphicsitem_cast<qnode_socket_t*>(item);
 
-          if (item_c && item_c != this) {
-
-              astate->tlog("CASTED!");
+          if (item_c && item_c != this && m_is_inp_socket != item_c->m_is_inp_socket) {
               //determine who is who
               //this is inp, target in out
+
 
               if (m_is_inp_socket && !item_c->m_is_inp_socket) {
                   m_connection->m_inp_socket = this;
                   m_connection->m_out_socket = item_c;
-                } else
-                if (!m_is_inp_socket && item_c->m_is_inp_socket) {
+                }
+
+              if (!m_is_inp_socket && item_c->m_is_inp_socket) {
                   m_connection->m_inp_socket = item_c;
                   m_connection->m_out_socket = this;
                 }
 
-              if (!m_connection->m_out_socket || !m_connection->m_inp_socket) {
-                  m_connection->setParentItem(nullptr);
-                  m_connection = nullptr;
-                } else {
-                  astate->tlog("m_connection->update_path(event->scenePos(), true)");
+              bool input_are_valid{false};
+              if (m_connection->m_inp_socket) {
+                  auto con_cnt = m_connection->m_inp_socket->connections_count();
+                  input_are_valid = con_cnt && *con_cnt == 0;
+                }
+
+              if (m_connection->m_out_socket && m_connection->m_inp_socket && input_are_valid) {
                   m_connection->update_path(event->scenePos(), true);
                   delete_item = false;
                   m_node->m_scene->m_connections.push_back(m_connection);
